chore(cubestore): Job processing refactoring
